backend/requirements.txt
aiosmtplib==5.0.0
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.0
bcrypt==5.0.0
beautifulsoup4==4.14.3
blinker==1.9.0
cachetools==6.2.3
certifi==2025.11.12
cffi==2.0.0
charset-normalizer==3.4.4
click==8.3.1
cryptography==46.0.3
curl_cffi==0.13.0
dnspython==2.8.0
ecdsa==0.19.1
email-validator==2.3.0
et_xmlfile==2.0.0
fastapi==0.124.2
fastapi-mail==1.6.0
frozendict==2.4.7
google-ai-generativelanguage==0.6.15
google-api-core==2.28.1
google-api-python-client==2.187.0
google-auth==2.43.0
google-auth-httplib2==0.2.1
google-generativeai==0.8.5
googleapis-common-protos==1.72.0
grpcio==1.76.0
grpcio-status==1.71.2
groq==0.8.0
h11==0.16.0
httpcore==1.0.9
httplib2==0.31.0
httpx==0.28.1
idna==3.11
iniconfig==2.3.0
Jinja2==3.1.6
joblib==1.5.2
MarkupSafe==3.0.3
multitasking==0.0.12
nltk==3.9.2
numpy==2.3.5
openpyxl==3.1.5
packaging==25.0
pandas==2.3.3
passlib==1.7.4
peewee==3.18.3
platformdirs==4.5.1
pluggy==1.6.0
proto-plus==1.26.1
protobuf==5.29.5
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.23
pydantic==2.12.5
pydantic-settings==2.12.0
pydantic_core==2.41.5
Pygments==2.19.2
pyparsing==3.2.5
pytest==9.0.2
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
python-jose==3.5.0
python-multipart==0.0.20
pytz==2025.2
regex==2025.11.3
requests==2.32.5
rsa==4.9.1
scikit-learn==1.8.0
scipy==1.16.3
six==1.17.0
soupsieve==2.8
SQLAlchemy==2.0.45
starlette==0.50.0
ta==0.11.0
textblob==0.19.0
threadpoolctl==3.6.0
tqdm==4.67.1
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.2
uritemplate==4.2.0
urllib3==2.6.2
uvicorn==0.38.0
websockets==15.0.1
yfinance==0.2.66

# Sentiment Analysis - VinSight v2
transformers>=4.30.0
torch>=2.0.0
